Refactoring to build everything around streams
